requirements.txt
azure-functions==1.7.2
selenium==3.141.0
urllib3==1.26.6
